requirements.txt
aiogram==3.20.0.post0
gspread==6.0.2
google-auth==2.29.0
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0 